+++ common/src/main/java/dev/juaanp/barehanded/config/ServerConfig.java
@@ -15,7 +15,7 @@ public class ServerConfig {
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
     private static final File FILE = Paths.get("config", Constants.MOD_ID + "-server.json").toFile();
 
-    public int configVersion = 6;
+    public int configVersion = 7;
 
     public double maxForce = 120.0;
     public double minDistance = 1.5;
@@ -29,8 +29,9 @@ public class ServerConfig {
     public boolean allowGrabbingSpawners = false;
 
     public boolean enableDistanceScroll = true;
+    public boolean dynamicScrollMaxDistance = true;
     public double scrollMinDistance = 1.0;
-    public double scrollMaxDistance = 12.0;
+    public double scrollMaxDistance = 256.0;
 
     public boolean enableRotation = true;
     public double rotationStabilization = 0.2;